add: extract tab count query
